consolodate generic halt and mkfs into abstract driver.
